test: avoid package growth for CLI seam
